-fixed some compiler warnings
Former-commit-id: c4caae0468abe0a2275b7e8ed625f0a2874ecf84

+++ VEngine/Code/Graphics/Vulkan/RenderGraph.h
@@ -326,7 +326,7 @@ namespace VEngine
 		// everything below needs to be reset before use //
 		///////////////////////////////////////////////////
 
-		std::vector<uint16_t> m_passSubresourceIndices;
+		std::vector<uint32_t> m_passSubresourceIndices;
 		std::vector<PassSubresources> m_passSubresources;
 
 		std::vector<ResourceDescription> m_resourceDescriptions;
